Refactor preposition learner tests into seperate file.
